[TWEAK] Move dependency injection extension classes into 'DependencyInjection.cs'.
https://github.com/codidact/authentication/pull/2/files/9ca8e7fd34ec78f95905388bcb3c885609b5294d#r380850907
